chore: rename cli directories
