+++ packages/aspida-mock/bin/index.js
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-require('../dist/lib/cli').run(process.argv.slice(2))
+require('../dist/cli').run(process.argv.slice(2))
