+++ packages/aspida-mock/src/cli/Template.ts
