+++ packages/aspida-mock/src/cli/listFiles.ts
